adding StaticTensor
